backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
scikit-learn
pandas
python-dotenv
requests
pytest
cryptography
numpy
rich